Switch from long to uint64_t for x_Handler
The uint64_t makes the size unambiguous. This change also fixes the Swift code
for RefMap to accomodate that.

+++ swift/CgoSwift/RefMap.swift
@@ -6,11 +6,11 @@ var globalRefMap = RefMap()
 class RefMap {
   // TODO: Use some sort of mixin for locking, if possible.
   private let q = dispatch_queue_create("RefMap", DISPATCH_QUEUE_SERIAL)
-  private var nextId: Int32 = 0
-  private var m: [Int32: Any] = [:]
+  private var nextId: UInt64 = 0
+  private var m: [UInt64: Any] = [:]
 
-  func put(x: Any) -> Int32 {
-    var h: Int32 = 0
+  func put(x: Any) -> UInt64 {
+    var h: UInt64 = 0
     dispatch_sync(q) {
       self.m[self.nextId] = x
       h = self.nextId
@@ -19,7 +19,7 @@ class RefMap {
     return h
   }
 
-  func get(h: Int32) -> Any? {
+  func get(h: UInt64) -> Any? {
     var x: Any? = nil
     dispatch_sync(q) {
       x = self.m[h]
@@ -27,7 +27,7 @@ class RefMap {
     return x
   }
 
-  func release(h: Int32) -> Any? {
+  func release(h: UInt64) -> Any? {
     var x: Any? = nil
     dispatch_sync(q) {
       x = self.m.removeValueForKey(h)
